adding error and checks in vote logic + various changes

+++ programs/solvote/src/constants.rs
@@ -1,2 +1,2 @@
 pub const MAX_OPTIONS_VALUE: usize = 5;
-pub const BASE_RATIO_VOTE: u8 = 100;
+pub const BASE_RATIO_VOTE: u16 = 100;
